// frontend/src/gen/api/src/models/NotificationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ESystemNotificationType } from './ESystemNotificationType';
import {
    ESystemNotificationTypeFromJSON,
    ESystemNotificationTypeFromJSONTyped,
    ESystemNotificationTypeToJSON,
} from './ESystemNotificationType';

/**
 * 
 * @export
 * @interface NotificationDto
 */
export interface NotificationDto {
    /**
     * 
     * @type {number}
     * @memberof NotificationDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof NotificationDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationDto
     */
    message?: string;
    /**
     * 
     * @type {ESystemNotificationType}
     * @memberof NotificationDto
     */
    notificationType?: ESystemNotificationType;
    /**
     * 
     * @type {Date}
     * @memberof NotificationDto
     */
    createdDate?: Date;
}

/**
 * Check if a given object implements the NotificationDto interface.
 */
export function instanceOfNotificationDto(value: object): boolean {
    return true;
}

export function NotificationDtoFromJSON(json: any): NotificationDto {
    return NotificationDtoFromJSONTyped(json, false);
}

export function NotificationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'message': json['message'] == null ? undefined : json['message'],
        'notificationType': json['notificationType'] == null ? undefined : ESystemNotificationTypeFromJSON(json['notificationType']),
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
    };
}

export function NotificationDtoToJSON(value?: NotificationDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'title': value['title'],
        'message': value['message'],
        'notificationType': ESystemNotificationTypeToJSON(value['notificationType']),
        'createdDate': value['createdDate'] == null ? undefined : ((value['createdDate']).toISOString()),
    };
}